+++ CHANGELOG.md
@@ -2,8 +2,9 @@ CHANGELOG
 ---------
 
 - **July.1.2016**: 1.5.0
+  - Added date/datetime default format code
   - Added multi-row header capability
-  - Added ability to style columns and ranges in xlsx using the `column_styles` and `custom_styles` options
+  - Added ability to style columns and ranges in xlsx using the `column_styles` and `range_styles` options
   - Adds `axlsx_styler` gem dependency
   - Refractor into smaller files
 - **May.16.2016**: 1.4.6